Refactoring DynamicBuffer based on divergence' review
